backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-core==1.4.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==41.0.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dnspython==2.4.2
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.106.0
fastjsonschema==2.19.1
fqdn==1.5.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
msal==1.26.0
msal-extensions==1.0.0
msrest==0.7.1
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.4
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
python-multipart==0.0.9